requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
fake-rpi==0.7.1
idna==2.10
numpy==1.19.2
opencv-python==4.4.0.44
paramiko==2.7.2
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pysftp==0.2.9
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
six==1.15.0
suntime==1.2.5
twilio==6.46.0
urllib3==1.25.11
